refactor(be): project structure
