+++ backend-nest/src/constants/constants.ts
@@ -1,1 +1,3 @@
 export const TTL = 20_000;
+
+export const GOLEMIO_API = "https://api.golemio.cz";
