+++ backend-nest/src/controllers/stop/stop.controller.ts
@@ -1,0 +1,4 @@
+import { Controller } from "@nestjs/common";
+
+@Controller("stop")
+export class StopController {}
